pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
    "ipykernel>=6.30.1",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.29",
    "langchain-text-splitters>=0.3.9",
    "langgraph>=0.6.4",
    "mcp[cli]>=1.13.0",
    "openai>=1.99.5",
    "python-dotenv>=1.1.1",
]

[tool.uv.workspace]
members = [
    "test_mcp/python-mcp-server",
]
